requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0
Pillow>=10.0.0
